+++ libs/eventually-express/src/ExpressApp.ts
@@ -28,6 +28,7 @@ import express, {
 import { Server } from "http";
 import * as swaggerUI from "swagger-ui-express";
 import { swagger } from "./swagger";
+import { OpenAPIV3_1 } from "openapi-types";
 
 const redoc = (title: string): string => `<!DOCTYPE html>
 <html>
@@ -68,7 +69,7 @@ export class ExpressApp extends AppBase {
   private _app = express();
   private _router = Router();
   private _server: Server;
-  private _swagger: swaggerUI.JsonObject;
+  private _swagger: OpenAPIV3_1.Document;
 
   public getSwagger(): any {
     return this._swagger;
@@ -327,6 +328,7 @@ export class ExpressApp extends AppBase {
     this._app.use(express.json());
     middleware && this._app.use(middleware);
     this._app.use(this._router);
+
     this._swagger = swagger(this);
     this._app.get("/swagger", (_, res: Response) => {
       res.json(this._swagger);
@@ -351,8 +353,34 @@ export class ExpressApp extends AppBase {
     this._app.get("/_endpoints", (_, res) => {
       res.json(this.endpoints);
     });
+    this._app.get("/_contracts", (_, res) => {
+      const schemas = this._swagger.components?.schemas;
+      const map = Object.keys(schemas).reduce(
+        (acc, name) => {
+          if (
+            schemas[name]?.properties?.name &&
+            schemas[name]?.properties?.id &&
+            schemas[name]?.properties?.stream &&
+            schemas[name]?.properties?.version &&
+            schemas[name]?.properties?.created &&
+            schemas[name]?.properties?.data
+          )
+            acc.events.push({
+              name,
+              payload: (
+                schemas[name].properties.data as OpenAPIV3_1.SchemaObject
+              ).properties,
+              schemaDescription: schemas[name].description
+            });
+          return acc;
+        },
+        { events: [] }
+      );
+      res.json(map);
+    });
+
     this._app.get("/_health", (_, res) => {
-      res.status(200).json({ status: "OK" });
+      res.status(200).json({ status: "OK", date: new Date().toISOString() });
     });
     this._app.get("/__killme", () => {
       this.log.info("red", "KILLME");
@@ -396,7 +424,8 @@ export class ExpressApp extends AppBase {
         redoc: "/redoc",
         rapidoc: "/rapidoc",
         health: "/_health",
-        endpoints: "/_endpoints"
+        endpoints: "/_endpoints",
+        contracts: "/_contracts"
       });
     });